move migration functions to their own file
